se hizo refact y se implemento la funcion de logout
