+++ frontend/src/app/(student)/chat/page.tsx
@@ -1,0 +1,6 @@
+import ChatInterface from '@/components/ChatInterface';
+import React from 'react';
+
+export default function StudentChatPage() {
+  return <ChatInterface />;
+}
