--- frontend/src/app/student/calendar/page.tsx
@@ -1,6 +1,0 @@
-import CalendarInterface from '../../components/CalendarInterface';
-import React from 'react';
-
-export default function CalendarPage() {
-    return <CalendarInterface />;
-}
