--- frontend/src/app/student/page.tsx
@@ -1,6 +1,0 @@
-import ChatInterface from '../components/ChatInterface';
-import React from 'react';
-
-export default function StudentChatPage(){
-  return <ChatInterface/>;
-}
