--- package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "gsap": "^3.13.0"
-  }
-}
